code/requirements.txt
backports.ssl-match-hostname==3.4.0.2
future==0.14.3
python-dockercloud==1.0.3
requests==2.7.0
schedule==0.3.2
six==1.9.0
websocket-client==0.30.0
wsgiref==0.1.2